examples/openai-agents/requirements.txt
openai-agents>=0.0.14
pydantic>=2.11.4
swarm-agent
openai>=1.0.0

# OpenTelemetry dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
